Pass `interval` and `langs` from command line arguments
